# patch: improve performance for lineage via prebuilding scope