--- a/cmd/const.go
+++ b/cmd/const.go
@@ -43,7 +43,7 @@ var (
 
 func variableOverridesMutator(variables []string) pipeline.PipelineMutator {
 	return func(ctx context.Context, p *pipeline.Pipeline) (*pipeline.Pipeline, error) {
-		var overrides = map[string]any{}
+		overrides := map[string]any{}
 		for _, variable := range variables {
 			parsed, err := parseVariable(variable)
 			if err != nil {