--- a/internal/requirements.txt
+++ b/internal/requirements.txt
@@ -1,1 +1,1 @@
-sqlglot[rs]==26.11.1
+sqlglot[rs]==26.19.0