This PR introduces two changes:
- it removes certain optimizations that are heavy in large queries
- it calculates and reuses a scope in lineage

based on my benchmarks, it improves the basic cases ~30% while allowing 4-5x improvement in some pathological queries.

```
Benchmark 1: before
  Time (mean ± σ):      3.630 s ±  0.022 s    [User: 0.054 s, System: 0.017 s]
  Range (min … max):    3.604 s …  3.680 s    10 runs
 
Benchmark 2: after
  Time (mean ± σ):     910.2 ms ±   9.8 ms    [User: 52.7 ms, System: 23.7 ms]
  Range (min … max):   902.0 ms … 932.8 ms    10 runs
 
Summary
  after ran
    3.99 ± 0.05 times faster than before
```

My understanding is that the larger the query, the bigger the difference. There still is a linear growth of the latency though, I couldn't solve that yet.